+++ ninja-core/src/main/java/ninja/utils/NoHttpBody.java
@@ -22,6 +22,8 @@ package ninja.utils;
  * It causes the ResultHandler to render no body, just the header. Useful
  * when issuing a redirect and no corresponding content should be shown.
  * 
+ * You might want to use Result.NO_HTTP_BODY as static shortcut.
+ * 
  * @author ra
  *
  */