+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,3 +1,9 @@
+Version X.X.X
+=============
+
+ * 2014-03-28 Added support for static Result.NO_HTTP_CONTENT. No need
+              to create new() class all the time. (ra)
+
 Version 3.1.3
 =============